WoodIntake/requirements.txt
streamlit
pandas
matplotlib
seaborn
numpy
requests
pillow
plotly
datetime
python-dotenv
shapely
figures
ifcopenshell
mathutils
streamlit-authenticator